backend/requirements.txt
flask
paho-mqtt
python-dotenv
flask-cors